requirements.txt
kivy
kivymd
buildozer
pillow
pygame

opencv-contrib-python-headless
numpy
scipy
scikit-image

mediapipe

easyocr
pytesseract
arabic-reshaper
python-bidi

gTTS
pyttsx3
SpeechRecognition
pyaudio

requests
tqdm
python-dotenv
joblib
appdirs
retry
cachetools
psutil